refactor(api): move contexts announcements and banner to new context communication
