--- api/src/announcements/README.md
@@ -1,11 +1,0 @@
-# Announcements context
-
-`announcements` is a context used to XXX.
-
-## Core entities
-
-XXX
-
-## Important notes
-
-XXX
