--- api/src/announcements/dependencies.json
@@ -1,1 +1,0 @@
-{ "name": "announcements", "dependsOn": ["shared"], "circular": "forbidden" }
