--- api/src/banner/README.md
@@ -1,11 +1,0 @@
-# Banner context
-
-`banner` is a context used to XXX.
-
-## Core entities
-
-XXX
-
-## Important notes
-
-XXX
