+++ api/src/communication/README.md
@@ -1,0 +1,11 @@
+# Communication context
+
+`communication` is a context used to XXX.
+
+## Core entities
+
+XXX
+
+## Important notes
+
+XXX
