+++ api/src/communication/announcements/infrastructure/serializers/jsonapi/announcement-serializer.js
