+++ api/src/communication/banner/domain/models/information-banner.js
