+++ api/src/communication/banner/domain/usecases/get-information-banner.js
