+++ api/src/communication/banner/domain/usecases/index.js
@@ -1,5 +1,5 @@
-import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
-import boundedContext from '../../dependencies.json' with { type: 'json' };
+import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
+import boundedContext from '../../../dependencies.json' with { type: 'json' };
 import * as informationBannerRepository from '../../infrastructure/repositories/information-banner-repository.js';
 
 /**
