+++ api/src/communication/banner/infrastructure/serializers/jsonapi/information-banner-serializer.js
