+++ api/tests/announcements/integration/infrastructure/repositories/announcement-repository_test.js
@@ -1,5 +1,5 @@
-import { Announcement } from '../../../../../src/announcements/domain/models/Announcement.js';
-import * as announcementRepository from '../../../../../src/announcements/infrastructure/repositories/announcement-repository.js';
+import { Announcement } from '../../../../../src/communication/announcements/domain/models/Announcement.js';
+import * as announcementRepository from '../../../../../src/communication/announcements/infrastructure/repositories/announcement-repository.js';
 import { announcementsStorage } from '../../../../../src/shared/infrastructure/key-value-storages/index.js';
 import { expect } from '../../../../test-helper.js';
 import { knex } from '../../../../tooling/databases.js';
