+++ api/tests/banner/integration/infrastructure/repositories/information-banner-repository_test.js
@@ -1,4 +1,4 @@
-import * as informationBannerRepository from '../../../../../src/banner/infrastructure/repositories/information-banner-repository.js';
+import * as informationBannerRepository from '../../../../../src/communication/banner/infrastructure/repositories/information-banner-repository.js';
 import { informationBannersStorage } from '../../../../../src/shared/infrastructure/key-value-storages/index.js';
 import { expect } from '../../../../test-helper.js';
 import { domainBuilder } from '../../../../tooling/domain-builder/domain-builder.js';
