+++ api/tests/banner/unit/domain/serializers/information-banner-serializer_test.js
@@ -1,4 +1,4 @@
-import { informationBannerSerializer } from '../../../../../src/banner/infrastructure/serializers/jsonapi/information-banner-serializer.js';
+import { informationBannerSerializer } from '../../../../../src/communication/banner/infrastructure/serializers/jsonapi/information-banner-serializer.js';
 import { expect } from '../../../../test-helper.js';
 import { domainBuilder } from '../../../../tooling/domain-builder/domain-builder.js';
